fix #2845 - cast param
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@16146 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ includes/modules/set_ids_by_url_parameters.php
@@ -28,7 +28,7 @@ if (isset($_GET['info'])) {
 if (isset($_GET['cat'])) {
   $site = explode('_', $_GET['cat']);
   $cID = $site[0];
-  $cID = str_replace('c', '', $cID);
+  $cID = (int)str_replace('c', '', $cID);
   $_GET['cPath'] = xtc_get_category_path($cID);
   unset($_GET['cat']);
 }
